Deduplicate test behaviors
